Extract DEM overlay assignment policies
